requirements.txt
beautifulsoup4==4.12.2
ffmpeg==1.4
python-dotenv==1.0.0
ffmpeg-python==0.2.0
pytube==15.0.0
requests==2.28.0
requests_html==0.10.0
selenium==4.9.0
python-dotenv==1.0.0
gdown===5.1.0